pyproject.toml
[tool.poetry]
name = "db-manager"
version = "0.1.0"
description = ""
authors = ["Андрей Лагунов <tobeplayed@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.12.1"
black = "^24.10.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
